(ftp): migrate connection logic to FtpCore
